mkdir handler (#8)
